Driver Service

+++ Server/src/main/java/org/example/server/Affectation/Repository/PermisRepository.java
@@ -3,5 +3,5 @@ package org.example.server.Affectation.Repository;
 import org.example.server.Affectation.Model.Permis;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PermisRepository extends CrudRepository<Permis,Integer> {
+public interface PermisRepository extends CrudRepository<Permis,Long> {
 }
